+++ src/lib/Components/Bidding/Screens/ConfirmBid.tsx
@@ -41,7 +41,6 @@ interface ConfirmBidProps extends ViewProperties {
 
 interface ConformBidState {
   pollCount: number
-  intervalToken: number
   conditionsOfSaleChecked: boolean
 }
 
@@ -65,7 +64,6 @@ const bidderPositionMutation = graphql`
 export class ConfirmBid extends React.Component<ConfirmBidProps, ConformBidState> {
   state = {
     pollCount: 0,
-    intervalToken: 0,
     conditionsOfSaleChecked: false,
   }
 
@@ -81,7 +79,7 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConformBidState
       onError: e => {
         // TODO catch error!
         // this.verifyAndShowBidResult(null, e)
-        console.log("error!", e, e.message)
+        console.error("error!", e, e.message)
       },
       mutation: bidderPositionMutation,
       variables: {
@@ -94,29 +92,27 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConformBidState
     })
   }
 
-  verifyBidPosition(results, errors) {
-    const status = results.createBidderPosition.result.status
-    if (!errors && status === "SUCCESS") {
-      // bidder position is created but not processed yet
-      const positionId = results.createBidderPosition.result.position.id
-      const query = `
+  queryForBidPosition(bidderPositionID: string) {
+    const query = `
         {
           me {
-            bidder_position(id: "${positionId}") {
+            bidder_position(id: "${bidderPositionID}") {
+              id
               processed_at
               is_active
             }
           }
         }
       `
-      const interval = setInterval(() => {
-        metaphysics({
-          query,
-        }).then(this.checkBidPosition.bind(this))
-      }, 1000)
-      this.setState({
-        intervalToken: interval,
-      })
+    return metaphysics({ query })
+  }
+
+  verifyBidPosition(results, errors) {
+    // TODO: Need to handle if the results object is empty, for example if errors occurred and no request was made
+    const status = results.createBidderPosition.result.status
+    if (!errors && status === "SUCCESS") {
+      const positionId = results.createBidderPosition.result.position.id
+      this.queryForBidPosition(positionId).then(this.checkBidPosition.bind(this))
     } else {
       const message_header = results.createBidderPosition.result.message_header
       const message_description_md = results.createBidderPosition.result.message_description_md
@@ -125,24 +121,24 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConformBidState
   }
 
   checkBidPosition(result) {
-    // TODO: move polling logic to a separate file https://github.com/artsy/emission/pull/1025#discussion_r185931915
     const bidderPosition = result.data.me.bidder_position
     if (bidderPosition.processed_at) {
-      clearInterval(this.state.intervalToken)
       if (bidderPosition.is_active) {
         // wining
         this.showBidResult(true, "SUCCESS")
       } else {
         // outbid
-        // TODO: show message (get from metaphysics?)
         this.showBidResult(false, "ERROR_BID_LOW")
       }
     } else {
-      // poll again
       if (this.state.pollCount > MAX_POLL_ATTEMPTS) {
-        clearInterval(this.state.intervalToken)
+        // TODO: Present error message to user.
+      } else {
+        setTimeout(() => {
+          this.queryForBidPosition(bidderPosition.id).then(this.checkBidPosition.bind(this))
+        }, 1000)
+        this.setState({ pollCount: this.state.pollCount + 1 })
       }
-      this.setState({ pollCount: this.state.pollCount + 1 })
     }
   }